[WFS provider] Move code related to XML Schema analysis to a new helper class and QgsOgcUtils, so it can be reused by OAPIF
