+++ python/PyQt6/core/auto_generated/qgsogcutils.sip.in
@@ -216,6 +216,7 @@ Creates an OGC expression XML element from the ``exp`` expression.
 
 
 
+
 };
