+++ src/providers/wfs/qgswfsprovider.h
@@ -187,18 +187,13 @@ class QgsWFSProvider final : public QgsVectorDataProvider
     //! Set if metadata retrieval has been canceled (typically download of the schema)
     bool mMetadataRetrievalCanceled = false;
 
-    bool readAttributesFromSchemaWithoutGMLAS( QDomDocument &schemaDoc, const QString &prefixedTypename, QString &geometryAttribute, QgsFields &fields, Qgis::WkbType &geomType, QString &errorMsg, bool &mayTryWithGMLAS );
-
-    bool readAttributesFromSchemaWithGMLAS( const QByteArray &response, const QString &prefixedTypename, QString &geometryAttribute, QgsFields &fields, Qgis::WkbType &geomType, bool &geometryMaybeMissing, QString &errorMsg );
-
   protected:
     //! String used to define a subset of the layer
     QString mSubsetString;
 
     //! Flag if provider is valid
     bool mValid = true;
-    //! Namespace URL of the server (comes from DescribeFeatureDocument)
-    QString mApplicationNamespace;
+
     //! Server capabilities for this layer (generated from capabilities document)
     Qgis::VectorProviderCapabilities mCapabilities;
     //! Fields of this typename. Might be different from mShared->mFields in case of SELECT
@@ -213,12 +208,6 @@ class QgsWFSProvider final : public QgsVectorDataProvider
      * the geometry attribute and the thematic attributes with their types.
     */
     bool describeFeatureType( QString &geometryAttribute, QgsFields &fields, Qgis::WkbType &geomType, bool &geometryMaybeMissing );
-
-    /**
-     * For a given typename, reads the name of the geometry attribute, the
-     * thematic attributes and their types from a dom document. Returns true in case of success.
-    */
-    bool readAttributesFromSchema( QDomDocument &schemaDoc, const QByteArray &response, bool singleLayerContext, const QString &prefixedTypename, QString &geometryAttribute, QgsFields &fields, Qgis::WkbType &geomType, bool &geometryMaybeMissing, QString &errorMsg );
 
     //helper methods for WFS-T
 
@@ -240,8 +229,7 @@ class QgsWFSProvider final : public QgsVectorDataProvider
     bool getCapabilities();
     //! Records provider error
     void handleException( const QDomDocument &serverResponse );
-    //! Converts DescribeFeatureType schema geometry property type to WKBType
-    Qgis::WkbType geomTypeFromPropertyType( const QString &attName, const QString &propType );
+
     //! Convert the value to its appropriate XML representation
     QString convertToXML( const QVariant &value );
