+++ src/providers/wfs/qgswfsshareddata.h
@@ -100,9 +100,6 @@ class QgsWFSSharedData : public QObject, public QgsBackgroundCachedSharedData
     friend class QgsWFSProvider;
     friend class QgsWFSSingleFeatureRequest;
 
-    //! Datasource URI
-    QgsWFSDataSourceURI mURI;
-
     //! WFS version to use. Comes from GetCapabilities response
     QString mWFSVersion;
 
@@ -114,12 +111,6 @@ class QgsWFSSharedData : public QObject, public QgsBackgroundCachedSharedData
 
     //! Map a field name to the pair (typename, fieldname) that describes its source field
     QMap<QString, QPair<QString, QString>> mMapFieldNameToSrcLayerNameFieldName;
-
-    //! Map a field name to the pair (xpath, isNestedContent)
-    QMap<QString, QPair<QString, bool>> mFieldNameToXPathAndIsNestedContentMap;
-
-    //! Map a namespace prefix to its URI
-    QMap<QString, QString> mNamespacePrefixToURIMap;
 
     //! Preferred HTTP method
     Qgis::HttpMethod mHttpMethod = Qgis::HttpMethod::Get;
@@ -168,8 +159,6 @@ class QgsWFSSharedData : public QObject, public QgsBackgroundCachedSharedData
     void invalidateCacheBaseUnderLock() override;
 
     bool supportsLimitedFeatureCountDownloads() const override { return !( mWFSVersion.startsWith( QLatin1String( "1.0" ) ) ); }
-
-    QString layerName() const override { return mURI.typeName(); }
 
     bool hasServerSideFilter() const override { return !mWFSFilter.isEmpty(); }
